+++ app/containers/wallet/dialogs/WalletRestoreDialogContainer.js
@@ -3,6 +3,7 @@ import React, { Component } from 'react';
 import { observer, inject } from 'mobx-react';
 import WalletRestoreDialog from '../../../components/wallet/WalletRestoreDialog';
 import type { InjectedDialogContainerProps } from '../../../types/injectedPropsType';
+import validWords from '../../../../lib/valid-words.en';
 
 @inject('stores', 'actions') @observer
 export default class WalletRestoreDialogContainer extends Component {
@@ -24,12 +25,12 @@ export default class WalletRestoreDialogContainer extends Component {
   };
 
   render() {
-    const { wallets } = this.props.stores;
-    const { restoreRequest } = wallets;
+    const { restoreRequest } = this.props.stores.wallets;
 
     return (
       <WalletRestoreDialog
         mnemonicValidator={mnemonic => this.props.stores.wallets.isValidMnemonic(mnemonic)}
+        suggestedMnemonics={validWords}
         isSubmitting={restoreRequest.isExecuting}
         onSubmit={this.onSubmit}
         onCancel={this.onCancel}